draft specific instrument Job models
